+++ RyuSocks/Commands/ClientCommand.cs
@@ -1,6 +1,7 @@
 using RyuSocks.Packets;
 using RyuSocks.Types;
 using System;
+using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
 namespace RyuSocks.Commands
@@ -8,6 +9,7 @@ namespace RyuSocks.Commands
     public abstract class ClientCommand : Command
     {
         protected readonly SocksClient Client;
+        protected readonly List<ProxyEndpoint> RemoteClients = [];
         public bool Accepted { get; protected set; }
         public bool Ready { get; protected set; }
 
@@ -15,6 +17,12 @@ namespace RyuSocks.Commands
         /// The endpoint used by the proxy server for this command.
         /// </summary>
         public ProxyEndpoint ServerEndpoint { get; protected set; }
+
+        /// <summary>
+        /// The list of remote client endpoints
+        /// which interact with the <see cref="ServerEndpoint"/> created by the proxy server.
+        /// </summary>
+        public IReadOnlyList<ProxyEndpoint> ClientEndpoints => RemoteClients;
 
         protected ClientCommand(SocksClient client, ProxyEndpoint proxyEndpoint) : base(proxyEndpoint)
         {